コントローラー修正
