+++ src/main/java/jp/co/aivick/app/controller/IndexController.java
@@ -6,9 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
 
-import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 
 import jp.co.aivick.app.entity.Recipe;
 import jp.co.aivick.app.service.RecipeService;
@@ -27,10 +25,4 @@ public class IndexController {
 		return "/index.html";
 	}
 
-	@RequestMapping("/json")
-	@ResponseBody
-	@JsonSerialize
-	public Recipe showModal(String id) {
-		return recipeService.findBy(id);
-	}
 }
